src/main/java/com/example/jpamapstruct/controller/EventController.java
package com.example.jpamapstruct.controller;

import com.example.jpamapstruct.dto.EventDto;
import com.example.jpamapstruct.entity.Event;
import com.example.jpamapstruct.mapper.EventMapper;
import com.example.jpamapstruct.service.EventService;
//import com.sun.tools.javac.util.DefinedBy.Api;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageImpl;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.data.rest.webmvc.ResourceNotFoundException;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;

@RequestMapping("/event")
@RestController
@Slf4j
//@Api("event")
public class EventController {
    private final EventService eventService;

    public EventController(EventService eventService) {
        this.eventService = eventService;
    }

    @PostMapping("/save")
    public ResponseEntity<Void> save(@RequestBody @Validated EventDto eventDto) {
        eventService.save(eventDto);
        return ResponseEntity.ok().build();
    }

    @GetMapping("/{id}")
    public ResponseEntity<EventDto> findById(@PathVariable("id") Long id) {
        EventDto event = eventService.findById(id);
        return ResponseEntity.ok(event);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
        Optional.ofNullable(eventService.findById(id)).orElseThrow(() -> {
            log.error("Unable to delete non-existent data！");
            return new ResourceNotFoundException();
        });
        eventService.deleteById(id);
        return ResponseEntity.ok().build();
    }

    @GetMapping("/page-query")
    public ResponseEntity<Page<EventDto>> pageQuery(EventDto eventDto, @PageableDefault(sort = "createAt", direction = Sort.Direction.DESC) Pageable pageable) {
        Page<EventDto> eventPage = eventService.findByCondition(eventDto, pageable);
        return ResponseEntity.ok(eventPage);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Void> update(@RequestBody @Validated EventDto eventDto, @PathVariable("id") Long id) {
        eventService.update(eventDto, id);
        return ResponseEntity.ok().build();
    }
}